# artifacts/api-server/src/routes/homework.ts
/**
 * /api/homework/* — teacher-style assignment CRUD.
 *
 * A homework session bundles a set of pages a student must work on by a
 * `dueDate`, each tagged "memorize" (new) or "revise" (existing). Grading
 * an item is a real recitation: the PATCH item handler ALSO writes to
 * page_progress and recitation_log, so the dashboard / streak / due-date
 * machinery in routes/progress.ts stays consistent.
 *
 * Status derivation is denormalized at read time:
 *   - completed → all items completed
 *   - overdue   → dueDate's calendar day has fully passed (see
 *                 isHomeworkOverdue — note the +1-day grace window)
 *   - active    → otherwise
 *
 * Undo of a recitation in routes/progress.ts will re-derive `completed`
 * for affected items in active sessions. See that file's
 * DELETE /progress/activity/:id handler.
 */
import { Router, type IRouter } from "express";
import { db, homeworkSessionsTable, homeworkItemsTable, pageProgressTable, recitationLogTable, ayahMistakesTable, ayahAttemptsTable } from "@workspace/db";
import { eq, and, sql, count, gte, inArray } from "drizzle-orm";
import {
  ListHomeworkResponse,
  CreateHomeworkBody,
  GetHomeworkParams,
  GetHomeworkResponse,
  GetHomeworkAyahsParams,
  GetHomeworkAyahsResponse,
  UpdateHomeworkParams,
  UpdateHomeworkBody,
  UpdateHomeworkResponse,
  DeleteHomeworkParams,
  UpdateHomeworkItemParams,
  UpdateHomeworkItemBody,
  UpdateHomeworkItemResponse,
} from "@workspace/api-zod";
import { ensurePageExists, getSettings, calculateDueDate, getDefaultPageName, getWeeklyReadCounts } from "../lib/progress-helpers";
import { requireAuth } from "../middlewares/requireAuth";
import { awardRecitationPoints } from "../lib/rewards";
import pageAyahsData from "../lib/page-ayahs.json" with { type: "json" };

const PAGE_AYAHS = pageAyahsData as Record<string, number[]>;

const router: IRouter = Router();

router.use(requireAuth);

/**
 * Homework is "overdue" only once the entire due date has passed — not the
 * instant the day begins. Due dates are stored as midnight UTC of the chosen
 * `YYYY-MM-DD`, so we treat the whole 24h window of that day as still active
 * and only flip to overdue from midnight UTC of the *next* day onwards.
 */
const ONE_DAY_MS = 24 * 60 * 60 * 1000;
function isHomeworkOverdue(dueDate: Date, now: Date): boolean {
  return now.getTime() >= dueDate.getTime() + ONE_DAY_MS;
}

router.get("/homework", async (req, res): Promise<void> => {
  const userId = req.userId!;
  const sessions = await db.select().from(homeworkSessionsTable)
    .where(eq(homeworkSessionsTable.userId, userId))
    .orderBy(homeworkSessionsTable.createdAt);
  const now = new Date();

  // Derive "completed" from the live pageProgressTable.quality (good or
  // excellent) rather than the historical `homeworkItemsTable.completed`
  // flag. The flag only flips when the user picks a quality from the
  // homework detail page; reciting the same page in the Reader updates
  // pageProgressTable but leaves the flag stale, which used to cause the
  // homework list progress bar to stay at 0/N. Joining here keeps the
  // list view consistent with what /homework/:id already returns.
  const itemCounts = await db
    .select({
      homeworkId: homeworkItemsTable.homeworkId,
      totalItems: count(),
      completedItems: count(
        sql`CASE WHEN ${pageProgressTable.quality} IN ('good', 'excellent') THEN 1 END`,
      ),
    })
    .from(homeworkItemsTable)
    .leftJoin(
      pageProgressTable,
      and(
        eq(pageProgressTable.userId, userId),
        eq(pageProgressTable.pageNumber, homeworkItemsTable.pageNumber),
      ),
    )
    .where(eq(homeworkItemsTable.userId, userId))
    .groupBy(homeworkItemsTable.homeworkId);

  const countsMap = new Map(itemCounts.map(c => [c.homeworkId, { total: Number(c.totalItems), completed: Number(c.completedItems) }]));

  const result = sessions.map(session => {
    const counts = countsMap.get(session.id) || { total: 0, completed: 0 };
    let status: string;
    if (counts.completed === counts.total && counts.total > 0) {
      status = "completed";
    } else if (isHomeworkOverdue(session.dueDate, now)) {
      status = "overdue";
    } else {
      status = "active";
    }

    return {
      id: session.id,
      title: session.title,
      dueDate: session.dueDate,
      createdAt: session.createdAt,
      completedItems: counts.completed,
      totalItems: counts.total,
      status,
    };
  });

  res.json(ListHomeworkResponse.parse(result));
});

router.post("/homework", async (req, res): Promise<void> => {
  const userId = req.userId!;
  const parsed = CreateHomeworkBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }

  const [session] = await db.insert(homeworkSessionsTable).values({
    userId,
    title: parsed.data.title,
    dueDate: new Date(parsed.data.dueDate),
    firstGlobalAyah: parsed.data.firstGlobalAyah ?? null,
    lastGlobalAyah: parsed.data.lastGlobalAyah ?? null,
  }).returning();

  const memorizePages = parsed.data.memorizePages || [];
  const revisePages = parsed.data.revisePages || [];

  for (const pageNumber of memorizePages) {
    await ensurePageExists(userId, pageNumber);
    await db.insert(homeworkItemsTable).values({
      userId,
      homeworkId: session.id,
      pageNumber,
      type: "memorize",
    });
  }

  for (const pageNumber of revisePages) {
    await ensurePageExists(userId, pageNumber);
    await db.insert(homeworkItemsTable).values({
      userId,
      homeworkId: session.id,
      pageNumber,
      type: "revise",
    });
  }

  const items = await db.select().from(homeworkItemsTable)
    .where(and(eq(homeworkItemsTable.userId, userId), eq(homeworkItemsTable.homeworkId, session.id)));

  res.status(201).json({
    id: session.id,
    title: session.title,
    dueDate: session.dueDate,
    createdAt: session.createdAt,
    completedItems: 0,
    totalItems: items.length,
    status: "active",
  });
});

router.get("/homework/:id", async (req, res): Promise<void> => {
  const userId = req.userId!;
  const params = GetHomeworkParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }

  const [session] = await db.select().from(homeworkSessionsTable)
    .where(and(eq(homeworkSessionsTable.userId, userId), eq(homeworkSessionsTable.id, params.data.id)));
  if (!session) {
    res.status(404).json({ error: "Homework session not found" });
    return;
  }

  const rows = await db
    .select({
      id: homeworkItemsTable.id,
      homeworkId: homeworkItemsTable.homeworkId,
      pageNumber: homeworkItemsTable.pageNumber,
      type: homeworkItemsTable.type,
      quality: pageProgressTable.quality,
      lastRecited: pageProgressTable.lastRecited,
      customName: pageProgressTable.customName,
    })
    .from(homeworkItemsTable)
    .leftJoin(
      pageProgressTable,
      and(
        eq(pageProgressTable.pageNumber, homeworkItemsTable.pageNumber),
        eq(pageProgressTable.userId, userId),
      ),
    )
    .where(and(eq(homeworkItemsTable.userId, userId), eq(homeworkItemsTable.homeworkId, session.id)))
    .orderBy(homeworkItemsTable.pageNumber);

  const pageNumbers = rows.map(r => r.pageNumber);
  const weekStart = new Date();
  weekStart.setDate(weekStart.getDate() - 6);
  weekStart.setHours(0, 0, 0, 0);

  // Weekly count includes BOTH quality recitations and explicit Telawa reads
  // (Khatmah + in-scope round-robin) in the trailing 7 days — "either counts".
  const weekCountMap = await getWeeklyReadCounts(userId, pageNumbers, weekStart);

  const detail = {
    id: session.id,
    title: session.title,
    dueDate: session.dueDate,
    createdAt: session.createdAt,
    firstGlobalAyah: session.firstGlobalAyah ?? null,
    lastGlobalAyah: session.lastGlobalAyah ?? null,
    items: rows.map(r => {
      const defaultName = getDefaultPageName(r.pageNumber);
      return {
        id: r.id,
        homeworkId: r.homeworkId,
        pageNumber: r.pageNumber,
        name: r.customName && r.customName.length > 0 ? r.customName : defaultName,
        customName: r.customName ?? null,
        type: r.type,
        completed: r.quality === "good" || r.quality === "excellent",
        quality: r.quality ?? null,
        completedAt: r.lastRecited ?? null,
        weekCount: weekCountMap.get(r.pageNumber) ?? 0,
      };
    }),
  };

  res.json(GetHomeworkResponse.parse(detail));
});

/**
 * GET /homework/:id/ayahs — the ayah-by-ayah view for a homework session.
 *
 * Enumerates every ayah on the homework's pages (via the static page→ayah
 * map) and annotates each with:
 *   - its current ACTIVE statuses (cleared / memorization / link) from
 *     ayah_mistakes where resolvedAt IS NULL,
 *   - `lastStatusAt` = the newest recitedAt among those active marks,
 *   - `weekAttemptCount` = number of ayah_attempts rows in the last 7 days.
 * `lastVisitedGlobalAyahNumber` = the ayah with the single most recent
 * active mark across the whole homework (the "last visited" highlight).
 *
 * Ayah text / surah name are intentionally NOT sent — the client resolves
 * them from its bundled quran dump (AyahIndex).
 */
router.get("/homework/:id/ayahs", async (req, res): Promise<void> => {
  const userId = req.userId!;
  const params = GetHomeworkAyahsParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }

  const [session] = await db.select().from(homeworkSessionsTable)
    .where(and(eq(homeworkSessionsTable.userId, userId), eq(homeworkSessionsTable.id, params.data.id)));
  if (!session) {
    res.status(404).json({ error: "Homework session not found" });
    return;
  }

  const items = await db
    .select({ pageNumber: homeworkItemsTable.pageNumber })
    .from(homeworkItemsTable)
    .where(and(eq(homeworkItemsTable.userId, userId), eq(homeworkItemsTable.homeworkId, session.id)));

  // Unique, sorted list of pages in this homework.
  const pageNumbers = Array.from(new Set(items.map(i => i.pageNumber))).sort((a, b) => a - b);

  // Build the ordered ayah list from the static page→ayah map.
  const ayahEntries: { globalAyahNumber: number; pageNumber: number }[] = [];
  for (const page of pageNumbers) {
    const ayahs = PAGE_AYAHS[String(page)] ?? [];
    for (const globalAyahNumber of ayahs) {
      ayahEntries.push({ globalAyahNumber, pageNumber: page });
    }
  }

  // Apply ayah-level boundary filter when the session has explicit bounds.
  // Pages always use the ceiling rule (whole page included); only the ayah
  // view respects the tight boundary.
  const firstBound = session.firstGlobalAyah;
  const lastBound = session.lastGlobalAyah;
  const filteredEntries =
    firstBound != null && lastBound != null
      ? ayahEntries.filter(e => e.globalAyahNumber >= firstBound && e.globalAyahNumber <= lastBound)
      : ayahEntries;

  if (filteredEntries.length === 0) {
    res.json(GetHomeworkAyahsResponse.parse({ lastVisitedGlobalAyahNumber: null, ayahs: [] }));
    return;
  }

  const globalNumbers = filteredEntries.map(e => e.globalAyahNumber);

  // Active per-ayah marks for these ayahs (resolvedAt IS NULL).
  const activeMarks = await db
    .select({
      globalAyahNumber: ayahMistakesTable.globalAyahNumber,
      mistakeType: ayahMistakesTable.mistakeType,
      recitedAt: ayahMistakesTable.recitedAt,
    })
    .from(ayahMistakesTable)
    .where(and(
      eq(ayahMistakesTable.userId, userId),
      inArray(ayahMistakesTable.globalAyahNumber, globalNumbers),
      sql`${ayahMistakesTable.resolvedAt} is null`,
    ));

  // Attempts in the trailing 7 days, grouped by ayah.
  const weekStart = new Date();
  weekStart.setDate(weekStart.getDate() - 6);
  weekStart.setHours(0, 0, 0, 0);

  const attemptRows = await db
    .select({
      globalAyahNumber: ayahAttemptsTable.globalAyahNumber,
      c: count(),
    })
    .from(ayahAttemptsTable)
    .where(and(
      eq(ayahAttemptsTable.userId, userId),
      inArray(ayahAttemptsTable.globalAyahNumber, globalNumbers),
      gte(ayahAttemptsTable.attemptedAt, weekStart),
    ))
    .groupBy(ayahAttemptsTable.globalAyahNumber);

  const attemptMap = new Map<number, number>();
  for (const r of attemptRows) attemptMap.set(r.globalAyahNumber, Number(r.c));

  // Fold active marks into per-ayah statuses + newest recitedAt.
  const statusMap = new Map<number, Set<string>>();
  const lastAtMap = new Map<number, Date>();
  let lastVisitedGlobalAyahNumber: number | null = null;
  let lastVisitedAt = 0;
  for (const m of activeMarks) {
    let set = statusMap.get(m.globalAyahNumber);
    if (!set) { set = new Set(); statusMap.set(m.globalAyahNumber, set); }
    set.add(m.mistakeType);
    const t = m.recitedAt ? m.recitedAt.getTime() : 0;
    const prev = lastAtMap.get(m.globalAyahNumber);
    if (m.recitedAt && (!prev || t > prev.getTime())) lastAtMap.set(m.globalAyahNumber, m.recitedAt);
    if (t > lastVisitedAt) { lastVisitedAt = t; lastVisitedGlobalAyahNumber = m.globalAyahNumber; }
  }

  const STATUS_ORDER = ["cleared", "memorization", "link"];
  const payload = {
    lastVisitedGlobalAyahNumber,
    ayahs: filteredEntries.map(e => {
      const set = statusMap.get(e.globalAyahNumber);
      const statuses = set ? STATUS_ORDER.filter(s => set.has(s)) : [];
      const lastAt = lastAtMap.get(e.globalAyahNumber);
      return {
        globalAyahNumber: e.globalAyahNumber,
        pageNumber: e.pageNumber,
        statuses,
        lastStatusAt: lastAt ? lastAt.toISOString() : null,
        weekAttemptCount: attemptMap.get(e.globalAyahNumber) ?? 0,
      };
    }),
  };

  res.json(GetHomeworkAyahsResponse.parse(payload));
});

router.patch("/homework/:id", async (req, res): Promise<void> => {
  const userId = req.userId!;
  const params = UpdateHomeworkParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }

  const parsed = UpdateHomeworkBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }

  const updateData: Record<string, unknown> = {};
  if (parsed.data.title !== undefined) updateData.title = parsed.data.title;
  if (parsed.data.dueDate !== undefined) updateData.dueDate = new Date(parsed.data.dueDate);
  if (parsed.data.firstGlobalAyah !== undefined) updateData.firstGlobalAyah = parsed.data.firstGlobalAyah ?? null;
  if (parsed.data.lastGlobalAyah !== undefined) updateData.lastGlobalAyah = parsed.data.lastGlobalAyah ?? null;

  // Confirm the session belongs to this user before touching items, even
  // when no session-level fields changed (the page-list fields below
  // mutate child rows independently and need the same authz check).
  const [existingSession] = await db
    .select()
    .from(homeworkSessionsTable)
    .where(and(eq(homeworkSessionsTable.userId, userId), eq(homeworkSessionsTable.id, params.data.id)));
  if (!existingSession) {
    res.status(404).json({ error: "Homework session not found" });
    return;
  }

  let updated = existingSession;
  if (Object.keys(updateData).length > 0) {
    const [row] = await db
      .update(homeworkSessionsTable)
      .set(updateData)
      .where(and(eq(homeworkSessionsTable.userId, userId), eq(homeworkSessionsTable.id, params.data.id)))
      .returning();
    if (!row) {
      res.status(404).json({ error: "Homework session not found" });
      return;
    }
    updated = row;
  }

  // Reconcile a single page-type's items against the desired list:
  //   - delete items whose pageNumber dropped out of the list (this only
  //     removes the homework association — the underlying pageProgress
  //     row, recitation log entries, and ayah marks all stay intact);
  //   - insert items for pageNumbers that weren't there before, calling
  //     ensurePageExists so the new pages have a pageProgress row;
  //   - leave surviving items untouched so their quality / completed
  //     state is preserved across edits.
  const reconcileItems = async (type: "memorize" | "revise", desiredPages: number[]): Promise<void> => {
    const desired = Array.from(new Set(desiredPages.filter(p => Number.isInteger(p) && p > 0)));
    const existing = await db
      .select()
      .from(homeworkItemsTable)
      .where(and(
        eq(homeworkItemsTable.userId, userId),
        eq(homeworkItemsTable.homeworkId, updated.id),
        eq(homeworkItemsTable.type, type),
      ));
    const existingPages = new Set(existing.map(i => i.pageNumber));
    const desiredSet = new Set(desired);

    const toDeleteIds = existing.filter(i => !desiredSet.has(i.pageNumber)).map(i => i.id);
    const toAdd = desired.filter(p => !existingPages.has(p));

    if (toDeleteIds.length > 0) {
      await db.delete(homeworkItemsTable).where(and(
        eq(homeworkItemsTable.userId, userId),
        inArray(homeworkItemsTable.id, toDeleteIds),
      ));
    }
    for (const pageNumber of toAdd) {
      await ensurePageExists(userId, pageNumber);
      await db.insert(homeworkItemsTable).values({
        userId,
        homeworkId: updated.id,
        pageNumber,
        type,
      });
    }
  };

  if (parsed.data.memorizePages !== undefined) {
    await reconcileItems("memorize", parsed.data.memorizePages);
  }
  if (parsed.data.revisePages !== undefined) {
    await reconcileItems("revise", parsed.data.revisePages);
  }

  const items = await db.select().from(homeworkItemsTable)
    .where(and(eq(homeworkItemsTable.userId, userId), eq(homeworkItemsTable.homeworkId, updated.id)));
  const completedItems = items.filter(i => i.completed).length;
  const now = new Date();

  let status: string;
  if (completedItems === items.length && items.length > 0) {
    status = "completed";
  } else if (isHomeworkOverdue(updated.dueDate, now)) {
    status = "overdue";
  } else {
    status = "active";
  }

  res.json(UpdateHomeworkResponse.parse({
    id: updated.id,
    title: updated.title,
    dueDate: updated.dueDate,
    createdAt: updated.createdAt,
    completedItems,
    totalItems: items.length,
    status,
  }));
});

router.delete("/homework/:id", async (req, res): Promise<void> => {
  const userId = req.userId!;
  const params = DeleteHomeworkParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }

  await db.delete(homeworkItemsTable)
    .where(and(eq(homeworkItemsTable.userId, userId), eq(homeworkItemsTable.homeworkId, params.data.id)));
  const [deleted] = await db.delete(homeworkSessionsTable)
    .where(and(eq(homeworkSessionsTable.userId, userId), eq(homeworkSessionsTable.id, params.data.id)))
    .returning();

  if (!deleted) {
    res.status(404).json({ error: "Homework session not found" });
    return;
  }

  res.sendStatus(204);
});

router.patch("/homework/:homeworkId/items/:itemId", async (req, res): Promise<void> => {
  const userId = req.userId!;
  const params = UpdateHomeworkItemParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }

  const parsed = UpdateHomeworkItemBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }

  const [item] = await db.select().from(homeworkItemsTable)
    .where(and(
      eq(homeworkItemsTable.userId, userId),
      eq(homeworkItemsTable.id, params.data.itemId),
      eq(homeworkItemsTable.homeworkId, params.data.homeworkId),
    ));

  if (!item) {
    res.status(404).json({ error: "Homework item not found" });
    return;
  }

  const updateData: Record<string, unknown> = {};
  if (parsed.data.completed !== undefined) {
    updateData.completed = parsed.data.completed;
    if (!parsed.data.completed && parsed.data.quality === undefined) {
      updateData.completedAt = null;
    }
  }
  if (parsed.data.quality !== undefined) {
    updateData.quality = parsed.data.quality;
    updateData.completedAt = new Date();
  }

  const [updated] = await db
    .update(homeworkItemsTable)
    .set(updateData)
    .where(and(eq(homeworkItemsTable.userId, userId), eq(homeworkItemsTable.id, params.data.itemId)))
    .returning();

  if (parsed.data.quality) {
    await ensurePageExists(userId, item.pageNumber);
    const settings = await getSettings(userId);
    const now = new Date();
    const dueDate = calculateDueDate(now, parsed.data.quality, settings);

    const [priorPage] = await db
      .select({ quality: pageProgressTable.quality })
      .from(pageProgressTable)
      .where(and(eq(pageProgressTable.userId, userId), eq(pageProgressTable.pageNumber, item.pageNumber)))
      .limit(1);

    await db
      .update(pageProgressTable)
      .set({
        quality: parsed.data.quality,
        lastRecited: now,
        dueDate,
        inScope: true,
      })
      .where(and(eq(pageProgressTable.userId, userId), eq(pageProgressTable.pageNumber, item.pageNumber)));

    await db.insert(recitationLogTable).values({
      userId,
      pageNumber: item.pageNumber,
      quality: parsed.data.quality,
      recitedAt: now,
      dueDate,
    });

    await awardRecitationPoints(userId, item.pageNumber, priorPage?.quality ?? null, parsed.data.quality, now);
  }

  const weekStart = new Date();
  weekStart.setDate(weekStart.getDate() - 6);
  weekStart.setHours(0, 0, 0, 0);
  const weekCountMap = await getWeeklyReadCounts(userId, [item.pageNumber], weekStart);
  const itemWeekCount = weekCountMap.get(item.pageNumber) ?? 0;

  const currentProgress = await db
    .select({
      quality: pageProgressTable.quality,
      lastRecited: pageProgressTable.lastRecited,
      customName: pageProgressTable.customName,
    })
    .from(pageProgressTable)
    .where(and(eq(pageProgressTable.userId, userId), eq(pageProgressTable.pageNumber, item.pageNumber)));
  const globalQuality = currentProgress[0]?.quality ?? updated.quality;
  const globalLastRecited = currentProgress[0]?.lastRecited ?? updated.completedAt;
  const customName = currentProgress[0]?.customName ?? null;
  const defaultName = getDefaultPageName(updated.pageNumber);

  res.json(UpdateHomeworkItemResponse.parse({
    id: updated.id,
    homeworkId: updated.homeworkId,
    pageNumber: updated.pageNumber,
    name: customName && customName.length > 0 ? customName : defaultName,
    customName,
    type: updated.type,
    completed: globalQuality === "good" || globalQuality === "excellent",
    quality: globalQuality,
    completedAt: globalLastRecited,
    weekCount: itemWeekCount,
  }));
});

export default router;