# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Quran Memorization Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /homework:
    get:
      operationId: listHomework
      tags: [homework]
      summary: List homework sessions
      responses:
        "200":
          description: Homework sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HomeworkSession"
    post:
      operationId: createHomework
      tags: [homework]
      summary: Create a homework session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHomeworkBody"
      responses:
        "201":
          description: Created homework session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HomeworkSession"
  /homework/{id}:
    get:
      operationId: getHomework
      tags: [homework]
      summary: Get homework session with items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Homework session detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HomeworkSessionDetail"
        "404":
          description: Not found
    patch:
      operationId: updateHomework
      tags: [homework]
      summary: Update homework session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateHomeworkBody"
      responses:
        "200":
          description: Updated homework session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HomeworkSession"
    delete:
      operationId: deleteHomework
      tags: [homework]
      summary: Delete homework session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /homework/{id}/ayahs:
    get:
      operationId: getHomeworkAyahs
      tags: [homework]
      summary: List every ayah on a homework's pages with its per-ayah status
      description: >
        Returns all ayahs belonging to the pages in this homework session
        (full scope), each annotated with its current active statuses
        (cleared / memorization / link), the timestamp of the most recent
        active status, and how many times it was attempted in the trailing
        7 days. `lastVisitedGlobalAyahNumber` is the ayah with the most
        recent active mark. Ayah text / surah name are resolved client-side
        from the bundled quran dump.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Homework ayah list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HomeworkAyahList"
        "404":
          description: Not found
  /homework/{homeworkId}/items/{itemId}:
    patch:
      operationId: updateHomeworkItem
      tags: [homework]
      summary: Update a homework item (mark page as done)
      parameters:
        - name: homeworkId
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateHomeworkItemBody"
      responses:
        "200":
          description: Updated homework item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HomeworkItem"
...
components:
  schemas:
...
    HomeworkSession:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        dueDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        completedItems:
          type: integer
        totalItems:
          type: integer
        status:
          type: string
          enum: [active, completed, overdue]
      required:
        - id
        - title
        - dueDate
        - createdAt
        - completedItems
        - totalItems
        - status
    CreateHomeworkBody:
      type: object
      properties:
        title:
          type: string
        dueDate:
          type: string
          format: date-time
        memorizePages:
          type: array
          items:
            type: integer
        revisePages:
          type: array
          items:
            type: integer
        firstGlobalAyah:
          type: integer
          nullable: true
          description: First global ayah number (1–6236) of this homework's scope. When set together with lastGlobalAyah, per-ayah views filter to only these ayahs; pages always use the ceiling rule.
        lastGlobalAyah:
          type: integer
          nullable: true
          description: Last global ayah number (1–6236) of this homework's scope.
      required:
        - title
        - dueDate
    UpdateHomeworkBody:
      type: object
      properties:
        title:
          type: string
        dueDate:
          type: string
          format: date-time
        memorizePages:
          type: array
          items:
            type: integer
          description: "If provided, fully replaces the set of memorize-type pages on this homework. Items whose pageNumber is no longer in the list are deleted; new pageNumbers are inserted. Surviving items keep their existing quality / completed state."
        revisePages:
          type: array
          items:
            type: integer
          description: "Same semantics as memorizePages, for revise-type items."
        firstGlobalAyah:
          type: integer
          nullable: true
          description: Update the homework's ayah-level scope start. Pass null to clear.
        lastGlobalAyah:
          type: integer
          nullable: true
          description: Update the homework's ayah-level scope end. Pass null to clear.
    HomeworkSessionDetail:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        dueDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        firstGlobalAyah:
          type: ["integer", "null"]
          description: First global ayah of this homework's scope, or null if not set.
        lastGlobalAyah:
          type: ["integer", "null"]
          description: Last global ayah of this homework's scope, or null if not set.
        items:
          type: array
          items:
            $ref: "#/components/schemas/HomeworkItem"
      required:
        - id
        - title
        - dueDate
        - createdAt
        - firstGlobalAyah
        - lastGlobalAyah
        - items
    HomeworkItem:
      type: object
      properties:
        id:
          type: integer
        homeworkId:
          type: integer
        pageNumber:
          type: integer
        name:
          type: string
        customName:
          type: ["string", "null"]
        type:
          type: string
          enum: [memorize, revise]
        completed:
          type: boolean
        quality:
          type: ["string", "null"]
          enum: [excellent, good, hard, relearn, null]
        completedAt:
          type: ["string", "null"]
          format: date-time
        weekCount:
          type: integer
          description: Number of times this page has been recited in the past 7 days (including today)
      required:
        - id
        - homeworkId
        - pageNumber
        - name
        - customName
        - type
        - completed
        - quality
        - completedAt
        - weekCount
    UpdateHomeworkItemBody:
      type: object
      properties:
        completed:
          type: boolean
        quality:
          type: string
          enum: [excellent, good, hard, relearn]
    HomeworkAyahEntry:
      type: object
      properties:
        globalAyahNumber:
          type: integer
        pageNumber:
          type: integer
        statuses:
          type: array
          description: Active marks on this ayah. Empty when the ayah has no active status.
          items:
            type: string
            enum: [cleared, memorization, link]
        lastStatusAt:
          type: ["string", "null"]
          format: date-time
          description: Timestamp of the most recent active status on this ayah, or null when none.
        weekAttemptCount:
          type: integer
          description: Number of times a status was set on this ayah in the trailing 7 days.
      required:
        - globalAyahNumber
        - pageNumber
        - statuses
        - lastStatusAt
        - weekAttemptCount
    HomeworkAyahList:
      type: object
      properties:
        lastVisitedGlobalAyahNumber:
          type: ["integer", "null"]
          description: Global ayah number of the most recently marked ayah in this homework, or null.
        ayahs:
          type: array
          items:
            $ref: "#/components/schemas/HomeworkAyahEntry"
      required:
        - lastVisitedGlobalAyahNumber
        - ayahs
...